+++ mayan/apps/document_states/tests/test_workflow_instance_api.py
@@ -7,7 +7,7 @@ from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
 from ..permissions import (
     permission_workflow_instance_transition,
-    permission_workflow_template_view
+    permission_workflow_template_view, permission_workflow_tools
 )
 
 from .literals import TEST_WORKFLOW_INSTANCE_LOG_ENTRY_EXTRA_DATA
@@ -476,6 +476,140 @@ class WorkflowInstanceAPIViewTestCase(
 
         response = self._request_test_workflow_instance_log_entry_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+
+class WorkflowInstanceLaunchAPIViewTestCase(
+    DocumentTestMixin, WorkflowTemplateTestMixin, BaseAPITestCase
+):
+    auto_upload_test_document = False
+
+    def setUp(self):
+        super().setUp()
+        self._create_test_workflow_template(
+            add_test_document_type=True, auto_launch=False
+        )
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_transition()
+        self._create_test_document_stub()
+
+    def _request_test_workflow_instance_launch_api_view(self):
+        return self.post(
+            viewname='rest_api:workflow-instance-launch', kwargs={
+                'document_id': self.test_document.pk
+            }, data={
+                'workflow_template_id': self._test_workflow_template.pk
+            }
+        )
+
+    def test_workflow_instance_api_view_no_permission(self):
+        test_document_workflow_instance_count = self.test_document.workflows.count()
+
+        self._clear_events()
+
+        response = self._request_test_workflow_instance_launch_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(
+            self.test_document.workflows.count(),
+            test_document_workflow_instance_count
+        )
+        self.assertNotEqual(
+            list(
+                self.test_document.workflows.values_list(
+                    'workflow', flat=True
+                )
+            ), [self._test_workflow_template.pk]
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_workflow_instance_api_view_with_document_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_workflow_tools
+        )
+
+        test_document_workflow_instance_count = self.test_document.workflows.count()
+
+        self._clear_events()
+
+        response = self._request_test_workflow_instance_launch_api_view()
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        self.assertEqual(
+            self.test_document.workflows.count(),
+            test_document_workflow_instance_count
+        )
+        self.assertNotEqual(
+            list(
+                self.test_document.workflows.values_list(
+                    'workflow', flat=True
+                )
+            ), [self._test_workflow_template.pk]
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_workflow_instance_api_view_with_workflow_template_access(self):
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_tools
+        )
+
+        test_document_workflow_instance_count = self.test_document.workflows.count()
+
+        self._clear_events()
+
+        response = self._request_test_workflow_instance_launch_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(
+            self.test_document.workflows.count(),
+            test_document_workflow_instance_count
+        )
+        self.assertNotEqual(
+            list(
+                self.test_document.workflows.values_list(
+                    'workflow', flat=True
+                )
+            ), [self._test_workflow_template.pk]
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_workflow_instance_api_view_with_full_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_workflow_tools
+        )
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_tools
+        )
+
+        test_document_workflow_instance_count = self.test_document.workflows.count()
+
+        self._clear_events()
+
+        response = self._request_test_workflow_instance_launch_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        self.assertEqual(
+            self.test_document.workflows.count(),
+            test_document_workflow_instance_count + 1
+        )
+        self.assertEqual(
+            list(
+                self.test_document.workflows.values_list(
+                    'workflow', flat=True
+                )
+            ), [self._test_workflow_template.pk]
+        )
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)